// src/app/models/product.model.ts
export interface Product {

    id?: string;
  
    code?: string;
  
    name?: string;
  
    description?: string;
  
    price?: number;
  
    quantity?: number;
  
    inventoryStatus?: string;
  
    category?: string;
  
    image?: string;
  
    rating?: number;
  
    cantidad?: number; // Added cantidad property
    material?: string; // Added material property
    observacion?: string; // Added observacion property
    fechaSolicitada?: string; // Added fechaSolicitada property
    paqueteTrabajo?: string; // Added paqueteTrabajo property
    tecnico?: string; // Added tecnico property
  
  }
  